+++ packages/linter-js-yaml/lib/linter-js-yaml.js
@@ -14,7 +14,7 @@ export default {
       items: {
         type: 'string',
       },
-      description: 'List of YAML custom tags.',
+      description: 'List of YAML custom tags, each optionally followed by a space and the node kind (scalar, mapping, or sequence).',
     },
   },
 
@@ -23,9 +23,10 @@ export default {
 
     this.subscriptions = new CompositeDisposable();
     this.subscriptions.add(atom.config.observe('linter-js-yaml.customTags', (customTags) => {
-      this.Schema = yaml.Schema.create(customTags.map(tag =>
-        new yaml.Type(tag, { kind: 'scalar' }),
-      ));
+      this.Schema = yaml.Schema.create(customTags.map((tag) => {
+        const typeInfo = tag.split(' ');
+        return new yaml.Type(typeInfo[0], { kind: typeInfo[1] || 'scalar' });
+      }));
     }));
   },
 
@@ -38,25 +39,34 @@ export default {
       grammarScopes: ['source.yaml', 'source.yml'],
       scope: 'file',
       name: 'Js-YAML',
-      lintOnFly: true,
+      lintsOnChange: true,
       lint: (TextEditor) => {
+        if (!atom.workspace.isTextEditor(TextEditor)) {
+          return null;
+        }
         const filePath = TextEditor.getPath();
+        if (!filePath) {
+          // Invalid path
+          return null;
+        }
         const fileText = TextEditor.getText();
 
         const messages = [];
-        const processMessage = (type, message) => {
-          let line = message.mark.line;
+        const processMessage = (severity, message) => {
+          let { line } = message.mark;
           // Workaround for https://github.com/nodeca/js-yaml/issues/218
           const maxLine = TextEditor.getLineCount() - 1;
           if (line > maxLine) {
             line = maxLine;
           }
-          const column = message.mark.column;
+          const { column } = message.mark;
           return {
-            type,
-            text: message.reason,
-            filePath,
-            range: helper.generateRange(TextEditor, line, column),
+            severity,
+            excerpt: message.reason,
+            location: {
+              file: filePath,
+              position: helper.generateRange(TextEditor, line, column),
+            },
           };
         };
 
@@ -65,11 +75,11 @@ export default {
             filename: path.basename(filePath),
             schema: this.Schema,
             onWarning: (warning) => {
-              messages.push(processMessage('Warning', warning));
+              messages.push(processMessage('warning', warning));
             },
           });
         } catch (error) {
-          messages.push(processMessage('Error', error));
+          messages.push(processMessage('error', error));
         }
 
         return messages;