+++ packages/linter/lib/ui-registry.js
@@ -5,8 +5,8 @@ import { ui as validateUI } from './validate'
 import type { Linter, UI, MessagesPatch } from './types'
 
 class UIRegistry {
-  providers: Set<UI>;
-  subscriptions: CompositeDisposable;
+  providers: Set<UI>
+  subscriptions: CompositeDisposable
 
   constructor() {
     this.providers = new Set()
@@ -23,6 +23,9 @@ class UIRegistry {
       provider.dispose()
       this.providers.delete(provider)
     }
+  }
+  getProviders(): Array<UI> {
+    return Array.from(this.providers)
   }
   render(messages: MessagesPatch) {
     this.providers.forEach(function(provider) {
@@ -45,4 +48,4 @@ class UIRegistry {
   }
 }
 
-module.exports = UIRegistry
+export default UIRegistry