+++ packages/aligner/CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.2.4 (2017-07-23)
+- Fix not able to align trailing comment
+
+## v1.2.3 (2017-06-22)
+- Fix config selector not matching language properly
+
 ## v1.2.2 (2017-06-14)
 - Prevent aligning when no editor is passing in (#69)