+++ packages/linter/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.3.0
+
+* Remove support for legacy linter APIs
+* Add a button to open developer console on Linter error messages
+* Include `description` in message key (when string rather than promise)
+* Fix a bug where linter messages (file-scoped) would disappear when same buffer is opened in two editors and one of them is closed
+
+## 2.2.0
+
+* `linter:debug` overhaul with more debug information
+* Add `linterName` support to messages API v2 (Thanks @hansonw from Facebook)
+
 ## 2.1.4
 
 * Hotfix release for a regression introduced in last version (sorry everyone)